Pivot to new version. Add simulateTransaction method
