+++ src/commonMain/kotlin/exchange.dydx.abacus/state/manager/V4StateManagerAdaptor+Transfer.kt
@@ -215,46 +215,25 @@ internal fun V4StateManagerAdaptor.simulateWithdrawal(
 ) {
     val payload = withdrawPayloadJson()
 
-    if (hasChainTransactionsV2Impl) {
-        transaction(
-            iListOf(Transaction(TransactionType.simulateWithdraw, payload)),
-            TargetChain.NOBLE,
-        ) { response ->
-            val error = parseTransactionResponse(response)
-            if (error != null) {
-                DebugLogger.error("simulateWithdrawal error: $error")
-                callback(null)
-                return@transaction
-            }
-
-            val result = parser.decodeJsonObject(response)
-            if (result != null) {
-                val amountMap = parser.asMap(parser.asList(result["amount"])?.firstOrNull())
-                val amount = parser.asDecimal(amountMap?.get("amount"))
-                val usdcAmount = amount?.div(Numeric.decimal.TEN.pow(decimals))
-                callback(usdcAmount)
-            } else {
-                callback(null)
-            }
+    simulateTransaction(
+        iListOf(Transaction(TransactionType.Withdraw, payload)),
+        TargetChain.NOBLE,
+    ) { response ->
+        val error = parseTransactionResponse(response)
+        if (error != null) {
+            DebugLogger.error("simulateWithdrawal error: $error")
+            callback(null)
+            return@simulateTransaction
         }
-    } else {
-        transaction(TransactionType.simulateWithdraw, payload) { response ->
-            val error = parseTransactionResponse(response)
-            if (error != null) {
-                DebugLogger.error("simulateWithdrawal error: $error")
-                callback(null)
-                return@transaction
-            }
 
-            val result = parser.decodeJsonObject(response)
-            if (result != null) {
-                val amountMap = parser.asMap(parser.asList(result["amount"])?.firstOrNull())
-                val amount = parser.asDecimal(amountMap?.get("amount"))
-                val usdcAmount = amount?.div(Numeric.decimal.TEN.pow(decimals))
-                callback(usdcAmount)
-            } else {
-                callback(null)
-            }
+        val result = parser.decodeJsonObject(response)
+        if (result != null) {
+            val amountMap = parser.asMap(parser.asList(result["amount"])?.firstOrNull())
+            val amount = parser.asDecimal(amountMap?.get("amount"))
+            val usdcAmount = amount?.div(Numeric.decimal.TEN.pow(decimals))
+            callback(usdcAmount)
+        } else {
+            callback(null)
         }
     }
 }
@@ -265,46 +244,25 @@ internal fun V4StateManagerAdaptor.simulateTransferNativeToken(
 ) {
     val payload = transferNativeTokenPayloadJson()
 
-    if (hasChainTransactionsV2Impl) {
-        transaction(
-            iListOf(Transaction(TransactionType.simulateTransferNativeToken, payload)),
-            TargetChain.DYDX,
-        ) { response ->
-            val error = parseTransactionResponse(response)
-            if (error != null) {
-                DebugLogger.error("simulateTransferNativeToken error: $error")
-                callback(null)
-                return@transaction
-            }
-
-            val result = parser.decodeJsonObject(response)
-            if (result != null) {
-                val amountMap = parser.asMap(parser.asList(result["amount"])?.firstOrNull())
-                val amount = parser.asDecimal(amountMap?.get("amount"))
-                val tokenAmount = amount?.div(Numeric.decimal.TEN.pow(decimals))
-                callback(tokenAmount)
-            } else {
-                callback(null)
-            }
+    simulateTransaction(
+        iListOf(Transaction(TransactionType.TransferNativeToken, payload)),
+        TargetChain.DYDX,
+    ) { response ->
+        val error = parseTransactionResponse(response)
+        if (error != null) {
+            DebugLogger.error("simulateTransferNativeToken error: $error")
+            callback(null)
+            return@simulateTransaction
         }
-    } else {
-        transaction(TransactionType.simulateTransferNativeToken, payload) { response ->
-            val error = parseTransactionResponse(response)
-            if (error != null) {
-                DebugLogger.error("simulateTransferNativeToken error: $error")
-                callback(null)
-                return@transaction
-            }
 
-            val result = parser.decodeJsonObject(response)
-            if (result != null) {
-                val amountMap = parser.asMap(parser.asList(result["amount"])?.firstOrNull())
-                val amount = parser.asDecimal(amountMap?.get("amount"))
-                val tokenAmount = amount?.div(Numeric.decimal.TEN.pow(decimals))
-                callback(tokenAmount)
-            } else {
-                callback(null)
-            }
+        val result = parser.decodeJsonObject(response)
+        if (result != null) {
+            val amountMap = parser.asMap(parser.asList(result["amount"])?.firstOrNull())
+            val amount = parser.asDecimal(amountMap?.get("amount"))
+            val tokenAmount = amount?.div(Numeric.decimal.TEN.pow(decimals))
+            callback(tokenAmount)
+        } else {
+            callback(null)
         }
     }
 }
@@ -600,27 +558,18 @@ internal fun V4StateManagerAdaptor.transferNobleBalance(amount: BigDecimal) {
                 val ibcPayload =
                     parser.asString(parser.value(json, "route.transactionRequest.data"))
                 if (ibcPayload != null) {
-                    if (hasChainTransactionsV2Impl) {
-                        transaction(
-                            iListOf(
-                                Transaction(
-                                    TransactionType.SendNobleIBC,
-                                    ibcPayload
-                                )
-                            ),
-                            TargetChain.NOBLE
-                        ) {
-                            val error = parseTransactionResponse(it)
-                            if (error != null) {
-                                DebugLogger.error("transferNobleBalance error: $error")
-                            }
-                        }
-                    } else {
-                        transaction(TransactionType.SendNobleIBC, ibcPayload) {
-                            val error = parseTransactionResponse(it)
-                            if (error != null) {
-                                DebugLogger.error("transferNobleBalance error: $error")
-                            }
+                    transaction(
+                        iListOf(
+                            Transaction(
+                                TransactionType.SendNobleIBC,
+                                ibcPayload
+                            )
+                        ),
+                        TargetChain.NOBLE
+                    ) {
+                        val error = parseTransactionResponse(it)
+                        if (error != null) {
+                            DebugLogger.error("transferNobleBalance error: $error")
                         }
                     }
                 }
@@ -697,39 +646,24 @@ internal fun V4StateManagerAdaptor.cctpToNoble(
                         ),
                     )
 
-                    if (hasChainTransactionsV2Impl) {
-                        transaction(
-                            iListOf(
-                                Transaction(
-                                    TransactionType.WithdrawToNobleIBC,
-                                    payload
-                                )
-                            ),
-                            TargetChain.DYDX
-                        ) {
-                            val error = parseTransactionResponse(it)
-                            if (error != null) {
-                                DebugLogger.error("withdrawToNobleIBC error: $error")
-                                send(error, callback)
-                            } else {
-                                pendingCctpWithdraw = CctpWithdrawState(
-                                    state?.input?.transfer?.requestPayload?.data,
-                                    callback,
-                                )
-                            }
-                        }
-                    } else {
-                        transaction(TransactionType.WithdrawToNobleIBC, payload) {
-                            val error = parseTransactionResponse(it)
-                            if (error != null) {
-                                DebugLogger.error("withdrawToNobleIBC error: $error")
-                                send(error, callback)
-                            } else {
-                                pendingCctpWithdraw = CctpWithdrawState(
-                                    state?.input?.transfer?.requestPayload?.data,
-                                    callback,
-                                )
-                            }
+                    transaction(
+                        iListOf(
+                            Transaction(
+                                TransactionType.WithdrawToNobleIBC,
+                                payload
+                            )
+                        ),
+                        TargetChain.DYDX
+                    ) {
+                        val error = parseTransactionResponse(it)
+                        if (error != null) {
+                            DebugLogger.error("withdrawToNobleIBC error: $error")
+                            send(error, callback)
+                        } else {
+                            pendingCctpWithdraw = CctpWithdrawState(
+                                state?.input?.transfer?.requestPayload?.data,
+                                callback,
+                            )
                         }
                     }
                 } else {
