+++ src/commonMain/kotlin/exchange.dydx.abacus/state/v2/supervisor/NetworkHelper.kt
@@ -2,7 +2,6 @@ package exchange.dydx.abacus.state.v2.supervisor
 
 import exchange.dydx.abacus.output.PerpetualState
 import exchange.dydx.abacus.output.UsageRestriction
-import exchange.dydx.abacus.protocols.DYDXChainTransactionsProtocolV2
 import exchange.dydx.abacus.protocols.DataNotificationProtocol
 import exchange.dydx.abacus.protocols.LocalTimerProtocol
 import exchange.dydx.abacus.protocols.ParserProtocol
@@ -55,7 +54,6 @@ class NetworkHelper(
     internal var lastValidatorCallTime: Instant? = null
     internal var lastIndexerCallTime: Instant? = null
     internal val jsonEncoder = JsonEncoder()
-    val hasChainTransactionsV2Impl: Boolean = ioImplementations.chain as? DYDXChainTransactionsProtocolV2 != null
 
     private var indexerRestriction: UsageRestriction? = null
         set(value) {
@@ -480,15 +478,15 @@ class NetworkHelper(
 
     @Throws(Exception::class)
     fun transaction(
-        transactionType: TransactionType,
-        paramsInJson: String?,
+        transactions: IList<Transaction>,
+        targetChain: TargetChain,
         callback: (response: String) -> Unit,
     ) {
         val transactionsImplementation = ioImplementations.chain
         if (transactionsImplementation === null) {
             throw Exception("chain is not DYDXChainTransactionsProtocol")
         }
-        transactionsImplementation.transaction(transactionType, paramsInJson) { response ->
+        transactionsImplementation.transaction(transactions, targetChain) { response ->
             if (response != null) {
                 val time = if (!response.contains("error")) {
                     Clock.System.now()
@@ -507,16 +505,16 @@ class NetworkHelper(
     }
 
     @Throws(Exception::class)
-    fun transaction(
+    fun simulateTransaction(
         transactions: IList<Transaction>,
         targetChain: TargetChain,
         callback: (response: String) -> Unit,
     ) {
-        val transactionsImplementation = ioImplementations.chain as? DYDXChainTransactionsProtocolV2
+        val transactionsImplementation = ioImplementations.chain
         if (transactionsImplementation === null) {
             throw Exception("chain is not DYDXChainTransactionsProtocol")
         }
-        transactionsImplementation.transactionV2(transactions, targetChain) { response ->
+        transactionsImplementation.simulateTransaction(transactions, targetChain) { response ->
             if (response != null) {
                 val time = if (!response.contains("error")) {
                     Clock.System.now()
