+++ src/commonMain/kotlin/exchange.dydx.abacus/state/v2/supervisor/SubaccountSupervisor.kt
@@ -32,8 +32,6 @@ import exchange.dydx.abacus.state.manager.PlaceOrderMarketInfo
 import exchange.dydx.abacus.state.manager.PlaceOrderRecord
 import exchange.dydx.abacus.state.manager.TransactionQueue
 import exchange.dydx.abacus.state.manager.TransactionParams
-import exchange.dydx.abacus.state.manager.TransactionParamsV2
-import exchange.dydx.abacus.state.manager.TransactionQueueV2
 import exchange.dydx.abacus.state.model.ClosePositionInputField
 import exchange.dydx.abacus.state.model.TradeInputField
 import exchange.dydx.abacus.state.model.TradingStateMachine
@@ -354,7 +352,6 @@ internal class SubaccountSupervisor(
     }
 
     val transactionQueue = TransactionQueue(helper::transaction)
-    val transactionQueueV2 = TransactionQueueV2(helper::transaction)
 
     private fun uiTrackingParmas(interval: Double): IMap<String, Any> {
         return iMapOf(
@@ -425,25 +422,13 @@ internal class SubaccountSupervisor(
         if (isShortTermOrder) {
             val submitTimeMs = Clock.System.now().toEpochMilliseconds().toDouble()
             val uiDelayTimeMs = submitTimeMs - uiClickTimeMs
-            if (helper.hasChainTransactionsV2Impl) {
-                helper.transaction(iListOf(Transaction(TransactionType.CancelOrder, string)), TargetChain.DYDX) {
-                        response -> transactionCallback(response, uiDelayTimeMs, submitTimeMs)
-                }
-            } else {
-                helper.transaction(TransactionType.CancelOrder, string) {
-                        response -> transactionCallback(response, uiDelayTimeMs, submitTimeMs)
-                }
+            helper.transaction(iListOf(Transaction(TransactionType.CancelOrder, string)), TargetChain.DYDX) {
+                    response -> transactionCallback(response, uiDelayTimeMs, submitTimeMs)
             }
         } else {
-            if (helper.hasChainTransactionsV2Impl) {
-                transactionQueueV2.enqueue(
-                    TransactionParamsV2(iListOf(Transaction(TransactionType.CancelOrder, string)), TargetChain.DYDX, transactionCallback)
-                )
-            } else {
-                transactionQueue.enqueue(
-                    TransactionParams(TransactionType.CancelOrder, string, transactionCallback, uiClickTimeMs)
-                )
-            }
+            transactionQueue.enqueue(
+                TransactionParams(iListOf(Transaction(TransactionType.CancelOrder, string)), TargetChain.DYDX, transactionCallback)
+            )
         }
     }
 
@@ -501,30 +486,13 @@ internal class SubaccountSupervisor(
             val submitTimeMs = Clock.System.now().toEpochMilliseconds().toDouble()
             val uiDelayTimeMs = submitTimeMs - uiClickTimeMs
 
-            if (helper.hasChainTransactionsV2Impl) {
-                helper.transaction(iListOf(Transaction(TransactionType.PlaceOrder, string)), TargetChain.DYDX) {
-                        response -> transactionCallback(response, uiDelayTimeMs, submitTimeMs)
-                }
-            } else {
-                helper.transaction(TransactionType.PlaceOrder, string) {
-                        response -> transactionCallback(response, uiDelayTimeMs, submitTimeMs)
-                }
+            helper.transaction(iListOf(Transaction(TransactionType.PlaceOrder, string)), TargetChain.DYDX) {
+                    response -> transactionCallback(response, uiDelayTimeMs, submitTimeMs)
             }
         } else {
-            if (helper.hasChainTransactionsV2Impl) {
-                transactionQueueV2.enqueue(
-                    TransactionParamsV2(iListOf(Transaction(TransactionType.PlaceOrder, string)), TargetChain.DYDX, transactionCallback),
-                )
-            } else {
-                transactionQueue.enqueue(
-                    TransactionParams(
-                        TransactionType.PlaceOrder,
-                        string,
-                        transactionCallback,
-                        uiClickTimeMs
-                    )
-                )
-            }
+            transactionQueue.enqueue(
+                TransactionParams(iListOf(Transaction(TransactionType.PlaceOrder, string)), TargetChain.DYDX, transactionCallback),
+            )
         }
 
         return payload
@@ -552,33 +520,7 @@ internal class SubaccountSupervisor(
 
         lastOrderClientId = null
 
-        if (helper.hasChainTransactionsV2Impl) {
-            helper.transaction(iListOf(Transaction(TransactionType.PlaceOrder, string)), TargetChain.DYDX) { response ->
-                val submitTimeMs = Clock.System.now().toEpochMilliseconds().toDouble()
-                val error = parseTransactionResponse(response)
-
-                if (error == null) {
-                    tracking(
-                        AnalyticsEvent.TradePlaceOrder.rawValue,
-                        ParsingHelper.merge(uiTrackingParmas(submitTimeMs - clickTimeMs), analyticsPayload)?.toIMap()
-                    )
-                    helper.ioImplementations.threading?.async(ThreadingType.abacus) {
-                        this.placeOrderRecords.add(
-                            PlaceOrderRecord(
-                                subaccountNumber,
-                                payload.clientId,
-                                submitTimeMs,
-                            )
-                        )
-                        lastOrderClientId = clientId
-                    }
-                }
-                helper.send(error, callback, payload)
-            }
-
-            return payload
-        }
-        helper.transaction(TransactionType.PlaceOrder, string) { response ->
+        helper.transaction(iListOf(Transaction(TransactionType.PlaceOrder, string)), TargetChain.DYDX) { response ->
             val submitTimeMs = Clock.System.now().toEpochMilliseconds().toDouble()
             val error = parseTransactionResponse(response)
 
@@ -602,6 +544,7 @@ internal class SubaccountSupervisor(
         }
 
         return payload
+
     }
 
     @Throws(Exception::class)
@@ -808,18 +751,10 @@ internal class SubaccountSupervisor(
         val string = Json.encodeToString(payload)
         val submitTimeInMilliseconds = Clock.System.now().toEpochMilliseconds().toDouble()
 
-        if (helper.hasChainTransactionsV2Impl) {
-            helper.transaction(iListOf(Transaction(TransactionType.Faucet, string)), TargetChain.DYDX) { response ->
-                val error =
-                    parseFaucetResponse(response, amount, submitTimeInMilliseconds)
-                helper.send(error, callback, payload)
-            }
-        } else {
-            helper.transaction(TransactionType.Faucet, string) { response ->
-                val error =
-                    parseFaucetResponse(response, amount, submitTimeInMilliseconds)
-                helper.send(error, callback, payload)
-            }
+        helper.transaction(iListOf(Transaction(TransactionType.Faucet, string)), TargetChain.DYDX) { response ->
+            val error =
+                parseFaucetResponse(response, amount, submitTimeInMilliseconds)
+            helper.send(error, callback, payload)
         }
     }
